browser/devtools/webide/test/sidebars/test_duplicate_import.html

<!DOCTYPE html>

<html>

  <head>
    <meta charset="utf8">
    <title></title>

    <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript" src="chrome://mochikit/content/chrome-harness.js"></script>
    <script type="application/javascript;version=1.8" src="../head.js"></script>
    <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  </head>

  <body>
    <script type="application/javascript;version=1.8">
      window.onload = function() {
        SimpleTest.waitForExplicitFinish();

        Task.spawn(function*() {
          let win = yield openWebIDE();
          let docProject = getProjectDocument(win);
          let packagedAppLocation = getTestFilePath("../app");
          let hostedAppManifest = TEST_BASE + "hosted_app.manifest";

          yield win.AppProjects.load();
          is(win.AppProjects.store.object.projects.length, 0, "IDB is empty");

          info("to call importPackagedApp(" + packagedAppLocation + ")");
          yield win.projectList.importPackagedApp(packagedAppLocation);
          yield waitForUpdate(win, "project-validated");
          yield nextTick();

          info("to call importHostedApp(" + hostedAppManifest + ")");
          yield win.projectList.importHostedApp(hostedAppManifest);
          yield waitForUpdate(win, "project-validated");
          yield nextTick();

          info("to call importPackagedApp(" + packagedAppLocation + ") again");
          yield win.projectList.importPackagedApp(packagedAppLocation);
          yield waitForUpdate(win, "project-validated");

          let project = win.AppManager.selectedProject;
          is(project.location, packagedAppLocation, "Correctly reselected existing packaged app.");
          yield nextTick();

          info("to call importHostedApp(" + hostedAppManifest + ") again");
          yield win.projectList.importHostedApp(hostedAppManifest);
          yield waitForUpdate(win, "project-validated");
          project = win.AppManager.selectedProject;
          is(project.location, hostedAppManifest, "Correctly reselected existing hosted app.");
          yield nextTick();

          let panelNode = docProject.querySelector("#project-panel");
          let items = panelNode.querySelectorAll(".panel-item");
          // 4 controls, + 2 projects
          is(items.length, 6, "6 projects in panel");
          is(items[3].querySelector("span").textContent, "A name (in app directory)", "Panel text is correct");
          is(items[4].querySelector("span").textContent, "hosted manifest name property", "Panel text is correct");

          yield closeWebIDE(win);

          yield removeAllProjects();

          SimpleTest.finish();
        }).then(null, e => {
          ok(false, "Exception: " + e);
          SimpleTest.finish();
        });
      }
    </script>
  </body>
</html>
